smem cleanup: move capacity helpers off legacy allocator, migrate test_quant
