+++ python/flydsl/utils/smem_allocator.py
@@ -230,42 +230,3 @@ class SmemAllocator:
         memref_type = T.memref(total_size, T.i8(), memory_space=lds_space())
         op = memref.get_global(memref_type, self.global_sym_name)
         return get_op_result_or_value(op)
-
-
-# ==============================================================================
-# Shared Memory Capacity Check
-# ==============================================================================
-
-SMEM_CAPACITY_MAP = {
-    # ===================== AMD CDNA Architectures (Data Center Compute Cards) =====================
-    # CDNA 3 (MI300 Series) - 64KB LDS per CU
-    "gfx942": 65536,  # MI300A / MI300X: 64KB LDS per CU
-    # CDNA 4 (MI350 Series) - 160KB LDS per CU (key upgrade for CDNA4)
-    "gfx950": 163840,  # MI300C / MI300X Enhanced Models: 64KB LDS per CU
-    "gfx1100": 65536,  # RDNA3 (Navi 31 / W7900): 64KB max per workgroup
-    "gfx1151": 65536,  # RDNA3.5: 64KB LDS per WGP
-    "gfx1201": 65536,  # RDNA4: 64KB LDS per WGP
-    # GFX1250 - 320KB LDS (WGP$ unified, 5 × 64KB segments)
-    "gfx1250": 327680,  # 320KB configurable as LDS
-}
-
-
-def check_smem_capacity(allocated_bytes: int, arch: str = None):
-    """
-    Checks if the allocated shared memory fits within the device capacity.
-    """
-    if arch is None:
-        # Try to detect arch from environment or FlyDSL context if possible
-        # For now, default to a safe limit or skip check if unknown
-        return
-
-    if arch in SMEM_CAPACITY_MAP:
-        limit = SMEM_CAPACITY_MAP[arch]
-        if allocated_bytes > limit:
-            raise RuntimeError(
-                f"Shared Memory Overflow: Requested {allocated_bytes} bytes, "
-                f"but device {arch} limit is {limit} bytes."
-            )
-    else:
-        # Unknown arch, maybe warn or skip
-        pass
